+++ R/label_hint.R
@@ -2,9 +2,7 @@
 #'
 #' This function inserts a label and optional hint.
 #' @param inputId The input slot that will be used to access the value
-#' @param label Display label for the control, or `NULL` for no label
-#' @param hint_input Display hint label for the control, or `NULL` for
-#' no hint label
+#' @inheritParams control_label_params
 #' @return a label hint HTML shiny tag object
 #' @family Govstyle feedback types
 #' @export
@@ -36,8 +34,8 @@ label_hint <- function(
   gov_label <- shiny::tags$div(
     class = "govuk-form-group",
     id = inputId,
-    shiny::tags$label(shiny::HTML(label), class = "govuk-label"),
-    shiny::tags$div(hint_input, class = "govuk-hint")
+    shiny::tags$label(as_govuk_html(label), class = "govuk-label"),
+    shiny::tags$div(as_govuk_html(hint_input), class = "govuk-hint")
   )
   attachDependency(gov_label)
 }